+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - #319: Override quick connect username when using user@domain
 - #283: Support for native PowerShell remoting as new protocol
 ### Changed
+- #1898: Fixed RDP desktop scaling not taking effect on remote connections
 - #1777: Cleaned up VisualStudio project structure
 - #1767: Turned about window into a simple popup form
 - #1766: Converted components check page into options page